srv/requirements.txt
falcon
pyyaml
structlog
gunicorn
cryptography
pymongo
pycrypto